\name{conditionalProbNspadesDriver}
\alias{conditionalProbNspadesDriver}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
conditionalProbNspadesDriver
}
\description{
driver to invoke conditionalProbNspades
}
\usage{
conditionalProbNspadesDriver()
}
\value{
returns a matrix of conditional probabilities
}
\author{
Barry Zeeberg
}
\examples{
x<-conditionalProbNspadesDriver()
}
