\name{cumulativeConditionalProbNspades}
\alias{cumulativeConditionalProbNspades}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
cumulativeConditionalProbNspades
}
\description{
compute cumulative probabilities from the matrix of conditional probabilities
}
\usage{
cumulativeConditionalProbNspades()
}
\value{
returns no value but has side effect of printing tables of cumulative probabilities
}
\author{
Barry Zeeberg
}
\examples{
x<-cumulativeConditionalProbNspades()
}
